move some logic to component
